% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpumetropolis-package.R
\docType{package}
\name{gpumetropolis-package}
\alias{gpumetropolis}
\alias{gpumetropolis-package}
\title{gpumetropolis: GPU-portable vendor-agnostic Metropolis-Hastings sampler}
\description{
Provides a batched random-walk Metropolis-Hastings sampler whose log-density
evaluation kernel is written to be portable across GPU vendors and CPU back
ends. This version ships the CPU reference sampler
(\code{\link[=metropolis_gaussian_mean]{metropolis_gaussian_mean()}}) and the distributional equivalence harness
(\code{\link[=rhat]{rhat()}}, \code{\link[=ess]{ess()}}, \code{\link[=ks_equivalence]{ks_equivalence()}}) against which later GPU versions are
checked.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pcbrom/gpumetropolis}
  \item Report bugs at \url{https://github.com/pcbrom/gpumetropolis/issues}
}

}
\author{
\strong{Maintainer}: Pedro Carvalho Brom \email{pcbrom@gmail.com} (\href{https://orcid.org/0000-0002-1288-7695}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Pedro Carvalho Brom \email{pcbrom@gmail.com} (\href{https://orcid.org/0000-0002-1288-7695}{ORCID}) [copyright holder]
}

}
\keyword{internal}
